[MakeMyIsland] add buttons for locations in Maldives
The underlying problem here is the target area is currently too small
for the geocoder to be useful: we're not really getting anything useful
back from the OSM geocoder within Fonadhoo. So I've manually located
the three villages on the island and added them as buttons... this is
really for interface because although one button for the whole island
is probably OK, basically it was just replicating the <h1> and wasn't
obvious (or sensible) that the user needed to hit a button at all..

If the project expands to more areas of the Maldives we can either take
those out or revert to the geocoder.

The same too-small-area logic makes the "locate me automatically" a
bit silly, so that's been dropped from the front page too.

Connected to mysociety/fixmystreet#863