[CuidoMiCiudad] Inherit improved email layouts from core
Recent changes in core (github.com/mysociety/fixmystreet/commit/dcc5926)
have improved the layout of FMS emails on narrow screens.

This commit removes a bunch of the outdated overrides that CMC had in
place, allowing the core files to show through. The only files left in
templates/email/cuidomiciudad are there because we need them to house
translated strings.